backend/surveys_service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
Flask-CORS
marshmallow-sqlalchemy
